V1.2

See merge request fa_sw/perihub!5

+++ api/app/support/fileHandler.py
@@ -2,8 +2,13 @@ import shutil
 import filecmp
 import paramiko
 import os
+import time
 
 class fileHandler(object):
+
+    def getRemotePath(Cluster):
+        
+        return './PeridigmJobs/apiModels/'
     
     def getRemoteModelPath(Cluster, username, ModelName):
         
@@ -16,6 +21,52 @@ class fileHandler(object):
     def getUserPath(Cluster, username, ModelName):
         
         return './PeridigmJobs/apiModels/' + username
+
+    def removeFolderIfOlder(path, days, recursive):
+        
+        now = time.time()
+        names = []
+        for foldername in os.listdir(path):
+            
+            folderPath =os.path.join(path, foldername)
+
+            if os.path.getmtime(folderPath) < now - days * 86400:
+                names.append(foldername)
+                shutil.rmtree(folderPath)
+            else:
+                if recursive:
+                    for subfoldername in os.listdir(folderPath):
+                        if os.path.getmtime(os.path.join(folderPath, subfoldername)) < now - days * 86400:
+                            names.append(os.path.join(foldername, subfoldername))
+                            shutil.rmtree(os.path.join(folderPath, subfoldername))
+        return names
+
+    def removeFolderIfOlderSftp(sftp, path, days, recursive):
+        
+        now = time.time()
+        for foldername in sftp.listdir(path):
+            
+            folderPath = os.path.join(path, foldername)
+
+            if os.path.getmtime(folderPath) < now - days * 86400:
+                fileHandler.removeAllFolderSftp(sftp, folderPath, True)
+            else:
+                if recursive:
+                    for subfoldername in sftp.listdir(folderPath):
+                        if os.path.getmtime(os.path.join(folderPath, subfoldername)) < now - days * 86400:
+                            fileHandler.removeAllFolderSftp(sftp, os.path.join(folderPath, subfoldername), True)
+
+    def removeAllFolderSftp(sftp, remotepath, recursive):
+
+        for filename in sftp.listdir(remotepath):
+            for subfilename in sftp.listdir(os.path.join(remotepath, filename)):
+                if recursive:
+                    for subSubfilename in sftp.listdir(os.path.join(remotepath, subfilename)):
+                        sftp.remove(os.path.join(remotepath, subSubfilename))
+                sftp.remove(os.path.join(remotepath, subfilename))
+            sftp.remove(os.path.join(remotepath, filename))
+        sftp.rmdir(remotepath)
+
 
     def copyModelToCluster(username, ModelName, Cluster):
         
@@ -30,6 +81,19 @@ class fileHandler(object):
             for root, dirs, files in os.walk(localpath):
                 if len(files)==0:
                     return ModelName + ' has not been created yet'
+                inputExist = False
+                meshExist = False
+                for name in files:
+                    if name.split('.')[-1]=='yaml':
+                        inputExist = True
+                    if name.split('.')[-1]=='txt' or name.split('.')[-1]=='e' :
+                        meshExist = True
+
+                if inputExist==False:
+                    return 'Inputfile of ' + ModelName + ' has not been created yet'
+
+                if meshExist==False:
+                    return 'Meshfile of ' + ModelName + ' has not been created yet'
 
                 for name in files:
                     shutil.copy(os.path.join(root,name), os.path.join(remotepath,name))
@@ -60,6 +124,20 @@ class fileHandler(object):
             for root, dirs, files in os.walk(localpath):
                 if len(files)==0:
                     return ModelName + ' has not been created yet'
+                inputExist = False
+                meshExist = False
+                for name in files:
+                    if name.split('.')[-1]=='yaml':
+                        inputExist = True
+                    if name.split('.')[-1]=='txt' or name.split('.')[-1]=='e':
+                        meshExist = True
+
+                if inputExist==False:
+                    return 'Inputfile of ' + ModelName + ' has not been created yet'
+
+                if meshExist==False:
+                    return 'Meshfile of ' + ModelName + ' has not been created yet'
+
                 for name in files:
                     sftp.put(os.path.join(root,name), name)
 
@@ -67,6 +145,23 @@ class fileHandler(object):
             ssh.close()
             
             return 'Success'
+
+    def copyFileToFromPeridigmContainer(username, ModelName, Filename, ToOrFrom):
+        
+        localpath = './Output/' + os.path.join(username, ModelName)
+        remotepath = './peridigmJobs/' + os.path.join(username, ModelName)
+        if not os.path.exists(remotepath):
+            os.makedirs(remotepath)
+            # os.chown(remotepath, 'test')
+        if not os.path.exists(localpath):
+            return ModelName + ' has not been created yet'
+
+        if ToOrFrom:
+            shutil.copy(os.path.join(localpath,Filename), os.path.join(remotepath,Filename))
+        else:
+            shutil.copy(os.path.join(remotepath,Filename), os.path.join(localpath,Filename))
+        
+        return 'Success'
 
     def copyResultsFromCluster(username, ModelName, Cluster, allData):
         resultpath = './Results/' + os.path.join(username, ModelName)
@@ -111,7 +206,7 @@ class fileHandler(object):
             ssh.close()
             return True
 
-    def sftpToCluster(Cluster, username):
+    def sftpToCluster(Cluster):
         
         if Cluster=='FA-Cluster':
             username='f_peridi'
